ipn/localapi: ensure watcher is installed before /watch-ipn-bus/ responds with 200
This change delays the first flush in the /watch-ipn-bus/ handler
until after the watcher has been successfully installed on the IPN
bus.

Without this, the endpoint returns a 200 almost immediatly, and
only then installs a watcher for IPN events.  This means there's a
small window where events could be missed by clients after calling
WatchIPNBus().

Clients could work around this by asking for an initial value,
reading it with Next(), and discarding it, but it seems safer to
make the backend guarantee that you won't miss events from the
moment WatchIPNBus() returns.

Signed-off-by: salman <salman@tailscale.com>